Implement virtual terminal buffer with ANSI codes
Using nix::pty lib and termion, the application creates a pty in which
to run the child application (e.g. local nethack, or ssh/dgl/nethack).
Keyboard input on the original pty is relayed to the new pty.
Raw output from the game is also forwarded to the real terminal.
However, a virtual terminal buffer is maintained and managed in memory
by the program.
This game screen will be used to investigate the inventory and game
stats, and also possibly to automate startscumming.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>